Change output layer in tabular example from 100 to 2
Since the output for our dataset can only be 2 values, 0 or 1,
the network simply learns to ignore 98 out of the 100 outputs.

I found it confusing that we'd end up with 100 outputs given that
this is a binary classification problem, so I figured updating the docs
would help others understand this as well.