A class- / protocol-oriented design for callbacks.
There are a lot of different ways to design a callback API. This change explores
one direction of having a well-typed API. There are a lot of different tradeoffs
and I'm especially interested in getting everyone's great ideas together.

If I understand correctly, the callbacks in the fast.ai python framework can
control the training loop, and even modify the model or training data. This thus
results in a variety of API design choices. One principle we adopt is that the
callbacks should be well typed, and take specific arguments and return well
defined responses. That said, there are a number of ways to actually define the
interface. To illustrate them, we pick the callback that can inspect the batch
of training data, and optionally modify it. Below are a few alternates:

 1. `func preprocessBatch(batch: Model.Input) -> Model.Input?`
   In this spelling, the callback takes the batch, and optionally returns it.
   This allows the callback to optionally modify the batch, as well as return
   `nil` to short-circuit the training loop.

2.  `func preprocessBatch(batch: Model.Input) -> BatchResult<Model.Input>`
   In this spelling, the callback takes the batch, and returns a value from the
   enum `BatchResult`, which can take values: `.continue(input: Model.Input)`,
   `.skip` which skips the batch and continues on to the next minibatch, and
   `.stop` which exits the training loop. This formulation gives a lot more
   expressiveness to the callbacks to control the training loop.

    Variant: add a `.continueUnmodified` so the callback can be clearer in the
    default case when it doesn't modify the training data.

3.  `func preprocessBatch(batch: Model.Input) -> Model.Input throws`
   In this spelling, the callback can throw in order to signal a non-default
   condition and/or control the training loop. We could define special values
   that the callback should throw.

All of the above require the callback formulations assume the batch is passed
by value. However, we could alternately use `inout`, as this simplifies the
type signatures for the return values. (e.g. No need to redundantly specify the
`Model.Input` type in the return value in addition to the parameter type.) Note
that `inout` is short for copy-in-copy-out (with optimizations), which is
exactly the semantics we want. We thus have a few alternatives for how to write
the APIs with `inout`:

4.  `func preprocessBatch(batch: inout Model.Input) -> bool`
   Return a boolean to continue or not. This has a couple disadvantages
   including not being able to halt the training loop (or skip the batch), but
   perhaps more importantly, it's confusing to the users of the API (both
   implementers as well as consumers of the API) whether `true` means continue
   or `true` means skip/halt.

5.  This final alternative is the one proposed in the PR, where the function is
    written: `func preprocessBatch(batch: inout Model.Input -> CallbackResult`
    where `CallbackResult` is an enum with the possible values of: `.continue`,
    `.stop`, and `.skip`.

That said, there are a number of open design questions that we should carefully
evaluate including (but not limited to):

 1. Would it make sense to group some of the callbacks together because they're
    often overridden together?
 2. Should this be a protocol instead of a class? (Note: we'd have to use
    associated types, which has implications on the ergonomics of usage.)
 3. How can we gracefully handle composition of multiple sets of callbacks? In
    the comments I propose chaining them together serially, which some
    reasonable way to aggregate the `CallbackResult`s.