Update 12b_lm_pretrain.ipynb
fixed named parameters:  The [__init__ for `RNNTrainer`](https://github.com/fastai/fastai_docs/blob/master/dev_course/dl2/exp/nb_12a.py#L145-L146
) has keyword arguments `alpha` and `beta` 

However, in the notebook trying to initialize the `RNNTrainer` class with α=2., β=1.  This ends up resulting in an error because keyword args do not match, so this is a minor fix that addresses this.