Attempt to fix source links for TypeDispatch properties (#170)
* * Unwrap TypeDispatch functions so get_function_source works for
`encodes` and `decodes` Transform methods.

`get_function_source` relies on `inspect.getsourcelines()`, which
expects some sort of callable object. However, `encodes` and `decodes`
are of type TypeDispatch, and `get_function_source` returns an empty
string. This, in turn, causes `show_doc` to generate an empty link.

When the notebook is exported to local/notebook/showdoc, it fixes the
source links for `Cuda.encodes` that were demonstrated during today's
code walk-thru in notebook `05_data_core`. The same thing would happen
for other transform encoding/decoding implementations such as those of
`OneHotEncode`.

Note that the source link for these methods points to a github revision,
whereas the links for other objects are generally resolved to
`nbviewer` URLs. I'm not sure if this is intentional or what the rules
are.

* * Check TypeDispatch class without importing 05 notebook.
* Unwrap TypeDispatch in get_source_link(), too.