Transform refactor docs
Reading code in notebooks is great to understand how things work but not always why things exist and why they are done the way they are. Jeremy has explained the rationale behind Transform well in his walk-thrus and I'm trying to distill it into the notebook.

The main idea here is to use the text to explain what we want from Transform and leave the details on how to use it in the code examples.